Move code out of MEVPDynamicsKernel and into IDynamicsKernel.
